+++ CHANGELOG.md
@@ -4,8 +4,15 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Fixed
+
+- Fix a stray translucent window appearing beside the player window at startup
+
 ## [0.1.0] - 2026-06-18
 
 Initial release.
 
+[Unreleased]: https://github.com/tobihagemann/somnio/compare/player-0.1.0...HEAD
 [0.1.0]: https://github.com/tobihagemann/somnio/releases/tag/player-0.1.0